Changed parser filename and updated it
